% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrw.R
\name{abrw}
\alias{abrw}
\title{The function implementing the network structured screening method proposed in this manuscripts.}
\usage{
abrw(list_amatrix, utility, list_alpha, max_it = 100, threshold = 0)
}
\arguments{
\item{list_amatrix}{list_amatrix is a list contains the matrix information you want. Every matrix is an p by p adjacency matrix here.}

\item{utility}{utility is the marginal screening meature vector with length p.}

\item{list_alpha}{is the tuning parameter in the proposed method. It shold have the same length as the number of matrices.}

\item{max_it}{max_it it the maximum of the iteration}

\item{threshold}{thresdhold is an feasible choice to deliminate the small value in the utility. Default 0 means use the whole utility vector.}
}
\description{
The function implementing the network structured screening method proposed in this manuscripts.
}
