% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_corrhat}
\alias{calculate_corrhat}
\title{calculate the conditional correlation among the X's}
\usage{
calculate_corrhat(X, betaMat, n, d)
}
\arguments{
\item{X}{The data matrix for the X}

\item{betaMat}{the beta matrix from lasso approach and its variantes. the betaMat can be estimated from the function SPMBgraphsqrtKeh: the sqrt lasso mathod is deployed.}

\item{n}{the sample nize.}

\item{d}{the dimension.}
}
\description{
calculate the conditional correlation among the X's
}
