% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_est.R
\name{calculate_Test_threshold}
\alias{calculate_Test_threshold}
\title{calculate the threshold of  Test statistics for the independence of X with control of FDR.}
\usage{
calculate_Test_threshold(Corr, N, P, alpha_fdr = 0.01)
}
\arguments{
\item{Corr}{The estimated conditional correlation among X}

\item{N}{The sample size N}

\item{P}{The dimension of X}

\item{alpha_fdr}{The pregiven FDR level}
}
\value{
A matrix of size p by b contains the conditional correlation among the X.
}
\description{
calculate the threshold of  Test statistics for the independence of X with control of FDR.
}
