% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_Testhat}
\alias{calculate_Testhat}
\title{calculate the Test statistics for the independence of X.}
\usage{
calculate_Testhat(X, betaMat, n, d)
}
\arguments{
\item{X}{The data matrix for the X}

\item{betaMat}{the beta matrix from lasso approach and its variantes. the betaMat can be estimated from the function SPMBgraphsqrtKeh: the sqrt lasso mathod is deployed.}

\item{n}{the sample nize. dim(X)}

\item{d}{the dimension. dim(X)}
}
\description{
calculate the Test statistics for the independence of X.
}
