% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cMBKRall}
\alias{cMBKRall}
\title{The proposed marginal screening method based on the BKR independence test}
\arguments{
\item{Ti}{is the vector of the ordered censored Time (ascent)}

\item{censor}{is the censor indicator for the Ti. 1 means no censor and 0 means censoring.}

\item{X}{is the complete covariates for the samples.}

\item{weight_C}{the inverse weight in the survival analysis. it is the output of the function KMestimate_CTime.}
}
\value{
the cBKRmeasure
}
\description{
The proposed marginal screening method based on the BKR independence test
}
