% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_est.R
\name{estimate_Corr}
\alias{estimate_Corr}
\title{calculate the conditional correlation among the X's}
\usage{
estimate_Corr(X)
}
\arguments{
\item{X}{The data matrix for the X}
}
\value{
A matrix of size p by b contains the conditional correlation among the X.
}
\description{
calculate the conditional correlation among the X's
}
