% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_est.R
\name{estimate_network}
\alias{estimate_network}
\title{Estimate the gaussian graphical model with FDR control.}
\usage{
estimate_network(Corr, thres)
}
\arguments{
\item{Corr}{the calculated conditional correlation of of the data X.}

\item{thres}{the calculated threshold for the test.}
}
\value{
A matrix of size p by b contains the 0-1 graph.
}
\description{
Estimate the gaussian graphical model with FDR control.
}
