% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{KMestimate_CTime}
\alias{KMestimate_CTime}
\title{The KM for censoring time}
\usage{
KMestimate_CTime(Ti, censor)
}
\arguments{
\item{Ti}{is the vector of the ordered censored Time (ascent)}

\item{censor}{is the censor indicator for the Ti. 1 means no censor and 0 means censoring.}
}
\value{
weight: The inverse weight of the C time. used in the cMBKRall function.
}
\description{
The KM for censoring time
}
