% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrw.R
\name{PageRankaug}
\alias{PageRankaug}
\title{PageRank Augmented screening method proposed by Wu., Zhu. and Feng.(2018)}
\usage{
PageRankaug(A, Orig_RM, d, r, tol = 1e-06, max_it = 100, threshold = 0)
}
\arguments{
\item{A}{the adjacency matrix with diagonal elements equal 0.}

\item{Orig_RM}{the original screening measure.}

\item{d}{the paramter balances the original ranking measure and the network information}

\item{r}{the size of the variable set including all active predictors}

\item{tol}{error for  convergence condition.}

\item{max_it}{the maximum of iteration}

\item{threshold}{make all marignal screening measure that are smaller than the threshold equals 0 to denoise.}
}
\value{
A matrix of size p by 1 contains the PageRank Augmented screening measure
}
\description{
The algorithm based on the effective power iteration method.
}
