% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPMBgraphsqrtKeh}
\alias{SPMBgraphsqrtKeh}
\title{sqrt lasso to calculate the betaMat for the X. actually part of the TIGER method}
\usage{
SPMBgraphsqrtKeh(data, lambda, nlambda, d)
}
\arguments{
\item{data}{The data matrix for the X}

\item{lambda}{sequence of tuning parameter.}

\item{nlambda}{the number of lambda sequence.}

\item{d}{the dimension of X.}
}
\description{
sqrt lasso to calculate the betaMat for the X. actually part of the TIGER method
}
